requirements.txt
django>=5.2.8
dotenv>=0.9.9
django-cors-headers~=4.9.0
djangorestframework~=3.16.1
djangorestframework-camel-case~=1.4.2
openai>=1.10.0
python-dotenv>=1.0.0
psycopg2-binary~=2.9.11
djangorestframework-simplejwt